remove cast

+++ examples/polybench_hls/adi.py
@@ -87,11 +87,10 @@ def adi(ttype, TSTEPS, N):
                 for j in range(1, N - 1):
                     p[i, j] = int(-c / ((a * p[i, j - 1] + b)))
                     # - (d) * (u[j, (i - 1)])
-                    q[i, j] = int(
-                        (1 + (2 * d)) * u[j, i]
-                        - (f * u[j, i + 1]
-                        - a * q[i, j - 1]
-                    ) / (a * p[i, j - 1] + b))
+                    q[i, j] = (1 + (2 * d)) * u[j, i]
+                    - (f * u[j, i + 1]
+                    - a * q[i, j - 1]
+                    ) / (a * p[i, j - 1] + b)
 
                 v[N - 1, i] = 1.0
                 for j_rev in range(N - 1):
